# api/openapi.yaml
openapi: 3.1.0
info:
  title: uspace-authority national API
  version: 0.1.0
  description: |
    The competent authority's system of the Georgian U-space rebuild.
    This file is the contract: the only source of handlers and types
    (`api/gen/`), and an endpoint that is not in it does not exist
    (spec `00 §7`). Paths are versioned under `/v1`; changes within a
    version are additive only.

    Every error is `application/problem+json` in the shape of the
    `Problem` schema (the lab's `schemas/common/problem/v1`, decision
    M28).

    Every operation carries `security` (an empty list only for the
    health paths) and the vendor extension `x-audit`, which names the
    `events` row the operation writes:

    - `x-audit: none` the operation writes no `events` row (health and
      public reads only);
    - `x-audit: {event_type: <name>, entity_type: <name>, purpose:
      required | optional | none}` the operation writes one `events`
      row of that type, inside the same transaction as its effect;
      `purpose: required` means the request is refused without a
      `purpose` (every PII read, CLAUDE.md rule 6).

    Work packages add their path groups below under a `# WP-k` comment
    and regenerate `api/gen/` with `scripts/generate.sh`.
servers:
  - url: /
...
paths:
...
  /v1/policy:
    get:
      operationId: getPolicy
      tags: [policy]
      summary: The active policy
      description: The one active policy version and its thresholds.
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: none
      responses:
        "200":
          description: The active policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createPolicy
      tags: [policy]
      summary: Create a new policy version
      description: |
        Validates every threshold (finite and positive, E-15; a refusal
        names each field) and stores it as the next version, inactive.
        Activation is a separate, audited step.
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: policy_created, entity_type: authority_policy, purpose: none}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyInput"
      responses:
        "201":
          description: The new version, inactive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        default:
          $ref: "#/components/responses/Problem"
  /v1/policy/{version}/activate:
    post:
      operationId: activatePolicy
      tags: [policy]
      summary: Activate a policy version
      description: |
        Makes `version` the one active policy and publishes it to the
        processes that follow it. Only a version newer than the active
        one can be activated (followers apply only a higher version); a
        rollback is a new version with the old values (409 otherwise).
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: policy_activated, entity_type: authority_policy, purpose: none}
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: The policy now active.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        default:
          $ref: "#/components/responses/Problem"
  /v1/audit/events:
    get:
      operationId: listAuditEvents
      tags: [audit]
      summary: Audit events, newest first
      description: |
        Every filter is optional; pages are cut by `before_id` (the
        `next_before_id` of the previous page). Reading the log is itself
        recorded as an `audit_events_viewed` event.
      security:
        - sessionCookie: []
      x-roles: [admin, auditor]
      x-audit: {event_type: audit_events_viewed, entity_type: events, purpose: optional}
      parameters:
        - {name: entity_type, in: query, schema: {type: string}}
        - {name: entity_id, in: query, schema: {type: string}}
        - {name: actor_id, in: query, schema: {type: string}}
        - {name: event_type, in: query, schema: {type: string}}
        - {name: from, in: query, description: Inclusive lower bound of ts., schema: {type: string, format: date-time}}
        - {name: to, in: query, description: Exclusive upper bound of ts., schema: {type: string, format: date-time}}
        - {name: before_id, in: query, schema: {type: integer, format: int64, minimum: 1}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 500, default: 100}}
        - {name: purpose, in: query, description: Why the log is read; recorded on the view event., schema: {type: string}}
      responses:
        "200":
          description: One page of events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventPage"
        default:
          $ref: "#/components/responses/Problem"

  # WP-2: console sign-in (password, then TOTP) and sessions. A session
  # is table A's session JWT (M20), presented as a bearer by the web BFF
  # from the uspace_session cookie (docs/runbooks/session-contract.md).
...
  /v1/auth/mfa:
    post:
      operationId: verifyMFA
      tags: [auth]
      summary: Second sign-in step (TOTP or a recovery code)
      description: |
        Exchanges the challenge and a TOTP code (or one recovery code)
        for a session. A code is accepted once; a challenge allows a
        bounded number of attempts and is used once. The first success
        after enrolment confirms it and returns the recovery codes,
        once. Every failure is an `mfa_refused` event; the session is a
        `session_started` event.
      security: []
      x-audit: {event_type: session_started, entity_type: session, purpose: none}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFARequest"
      responses:
        "200":
          description: The session.
          headers:
            Cache-Control:
              schema: {type: string, enum: [no-store]}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionIssued"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Problem"
  /v1/auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: The session of the caller
      security:
        - sessionCookie: []
      x-session: any
      x-audit: none
      responses:
        "200":
          description: The session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        default:
          $ref: "#/components/responses/Problem"
...
  /v1/users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Console accounts
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: none
      responses:
        "200":
          description: Every account, by username.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createUser
      tags: [users]
      summary: Create an account
      description: |
        The account signs in with the password given here and enrols
        TOTP at its first sign-in.
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: user_created, entity_type: user, purpose: none}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: The account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Problem"
  /v1/users/{user_id}:
    parameters:
      - {name: user_id, in: path, required: true, schema: {type: string, maxLength: 32}}
    get:
      operationId: getUser
      tags: [users]
      summary: One account
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: none
      responses:
        "200":
          description: The account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Problem"
  /v1/users/{user_id}/roles:
    parameters:
      - {name: user_id, in: path, required: true, schema: {type: string, maxLength: 32}}
    put:
      operationId: setUserRoles
      tags: [users]
      summary: Replace the roles of an account
      description: Revokes the sessions of the account, so the new roles apply at the next sign-in. The last active admin cannot lose the role (409).
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: user_roles_changed, entity_type: user, purpose: none}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RolesInput"
      responses:
        "200":
          description: The account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Problem"
  /v1/users/{user_id}/disable:
    parameters:
      - {name: user_id, in: path, required: true, schema: {type: string, maxLength: 32}}
    post:
      operationId: disableUser
      tags: [users]
      summary: Disable an account
      description: Revokes its sessions; it can no longer sign in. The last active admin cannot be disabled (409).
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: user_disabled, entity_type: user, purpose: none}
      responses:
        "200":
          description: The account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Problem"
  /v1/users/{user_id}/enable:
    parameters:
      - {name: user_id, in: path, required: true, schema: {type: string, maxLength: 32}}
    post:
      operationId: enableUser
      tags: [users]
      summary: Enable a disabled account
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: user_enabled, entity_type: user, purpose: none}
      responses:
        "200":
          description: The account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Problem"
  /v1/users/{user_id}/mfa/reset:
    parameters:
      - {name: user_id, in: path, required: true, schema: {type: string, maxLength: 32}}
    post:
      operationId: resetUserMFA
      tags: [users]
      summary: Reset the TOTP of an account
      description: Deletes the TOTP secret and the recovery codes and revokes the sessions; the account enrols again at its next sign-in.
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: user_mfa_reset, entity_type: user, purpose: none}
      responses:
        "200":
          description: The account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Problem"
  /v1/users/{user_id}/mfa/unlock:
    parameters:
      - {name: user_id, in: path, required: true, schema: {type: string, maxLength: 32}}
    post:
      operationId: unlockUserMFA
      tags: [users]
      summary: Clear the MFA lock of an account
      description: |
        Wrong TOTP or recovery codes count against the account across
        challenges and addresses (NIST SP 800-63B 5.2.2): from
        `MFA_LOCKOUT_AFTER` failures the account is locked for a time
        that doubles with each failure, and at `MFA_HARD_LOCK_AFTER`
        (at most 100) until an admin calls this. A success clears the
        count.
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: user_mfa_unlocked, entity_type: user, purpose: none}
      responses:
        "200":
          description: The account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          $ref: "#/components/responses/Problem"
  /v1/users/{user_id}/sessions/revoke:
    parameters:
      - {name: user_id, in: path, required: true, schema: {type: string, maxLength: 32}}
    post:
      operationId: revokeUserSessions
      tags: [users]
      summary: Revoke every session of an account
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: session_revoked, entity_type: session, purpose: none}
      responses:
        "200":
          description: How many sessions were revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revoked"
        default:
          $ref: "#/components/responses/Problem"

  # WP-2: the ecosystem token service. /oauth/token and /.well-known/*
  # are public (the client authenticates in the body); /v1/oauth/* is
  # for admins. Every issuance and every refusal is an events row
  # (token_issued, token_refused; 06 §3).
  /oauth/token:
    post:
      operationId: requestToken
      tags: [tokens]
      summary: Client credentials grant
      description: |
        RFC 6749 §4.4 client credentials. The client authenticates with
        `client_secret_post` (`client_id`, `client_secret`) or
        `private_key_jwt` (RFC 7523: `client_assertion_type` =
        `urn:ietf:params:oauth:client-assertion-type:jwt-bearer`,
        `client_assertion` an RS256 JWT with `iss` = `sub` = the client
        id, `aud` = this issuer or its token endpoint, `exp` at most
        5 min ahead and a `jti` used once). Parameters in the query
        string are refused. `scope` names catalogue scopes (WP-2 table
        B), a subset of the client's. `audience`, or one RFC 8707
        `resource`, names the target: a host or the target's base URL,
        reduced to its host, which becomes `aud` (M18). For national
        scopes the host must be on the client's audience list; for
        `utm.*` and `rid.*` any host is accepted. One audience per
        token. Errors carry RFC 6749 `error` and `error_description`
        beside the problem members. Issuance is rate-limited per
        client (429 with `Retry-After`). Every refusal is a
        `token_refused` event.
      security: []
      x-audit: {event_type: token_issued, entity_type: oauth_client, purpose: none}
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: An ecosystem machine token (table A).
          headers:
            Cache-Control:
              schema: {type: string, enum: [no-store]}
            Pragma:
              schema: {type: string, enum: [no-cache]}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/OAuthError"
        "401":
          $ref: "#/components/responses/OAuthError"
        "429":
          description: The client's issuance budget is spent.
          headers:
            Retry-After:
              schema: {type: integer, minimum: 1}
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/OAuthProblem"
        default:
          $ref: "#/components/responses/Problem"
  /.well-known/jwks.json:
    get:
      operationId: getJWKS
      tags: [tokens]
      summary: The issuer's public keys
      description: |
        Every RS256 key a verifier may meet (`use: sig`, distinguished by
        `kid`): the active token-signing key, retired token-signing keys
        until their retirement plus the JWKS cache TTL (24 h), and this
        system's publication-signing key of the detached
        `X-JWS-Signature` (M26). Cacheable for 300 s.
      security: []
      x-audit: none
      responses:
        "200":
          description: The JWKS.
          headers:
            Cache-Control:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWKS"
        default:
          $ref: "#/components/responses/Problem"
  /.well-known/openid-configuration:
    get:
      operationId: getIssuerMetadata
      tags: [tokens]
      summary: Issuer metadata
      description: |
        The issuer, its JWKS URL and token endpoint, the grant type and
        the client authentication methods. This is an OAuth2 issuer, not
        an OpenID provider: there is no authorisation endpoint.
      security: []
      x-audit: none
      responses:
        "200":
          description: The metadata.
          headers:
            Cache-Control:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuerMetadata"
        default:
          $ref: "#/components/responses/Problem"
  /v1/oauth/clients:
    get:
      operationId: listOAuthClients
      tags: [oauth-admin]
      summary: Registered clients
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: none
      responses:
        "200":
          description: Every client, by client id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthClientList"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createOAuthClient
      tags: [oauth-admin]
      summary: Register a client
      description: |
        One client per calling system (M24). Scopes must be grantable
        to this client (in the catalogue, not reserved, not of a USSP's
        issuer, `dp.observe` for `lab-01` only); audiences are hosts.
        For `client_secret_post` the secret is generated here and
        returned once; only its argon2id hash is kept. For
        `private_key_jwt` the client's public JWKS is given (RSA keys of
        at least 2048 bits, no private members).
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: oauth_client_created, entity_type: oauth_client, purpose: none}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OAuthClientInput"
      responses:
        "201":
          description: The client, and its secret once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthClientCreated"
        default:
          $ref: "#/components/responses/Problem"
  /v1/oauth/clients/{client_id}:
    parameters:
      - {name: client_id, in: path, required: true, schema: {type: string, maxLength: 64}}
    get:
      operationId: getOAuthClient
      tags: [oauth-admin]
      summary: One client
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: none
      responses:
        "200":
          description: The client.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthClient"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateOAuthClient
      tags: [oauth-admin]
      summary: Change a client's status, scopes, audiences or note
      description: |
        A suspended or revoked client gets no new token; tokens already
        issued expire within their TTL (at most 1 h).
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: oauth_client_updated, entity_type: oauth_client, purpose: none}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OAuthClientPatch"
      responses:
        "200":
          description: The client as changed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthClient"
        default:
          $ref: "#/components/responses/Problem"
  /v1/oauth/keys:
    get:
      operationId: listSigningKeys
      tags: [oauth-admin]
      summary: Signing keys and their state
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: none
      responses:
        "200":
          description: Every registered key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SigningKeyList"
        default:
          $ref: "#/components/responses/Problem"
  /v1/oauth/keys/{kid}/compromised:
    parameters:
      - {name: kid, in: path, required: true, schema: {type: string, maxLength: 128}}
    post:
      operationId: compromiseSigningKey
      tags: [oauth-admin]
      summary: Drop a compromised key now
      description: |
        The emergency retirement: the key leaves the JWKS at once (no
        24 h overlap), never signs again and is never a rotation
        candidate. An active token key is replaced by the next
        candidate; without one this issuer stops issuing until a key is
        added. One admin suffices; the act is a `signing_key_compromised`
        event with the reason. Other replicas follow within
        `KEY_REFRESH_S`; verifiers drop the key when their JWKS cache
        refreshes.
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: signing_key_compromised, entity_type: signing_key, purpose: none}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: {type: string, maxLength: 1000}
      responses:
        "200":
          description: The key as marked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SigningKey"
        default:
          $ref: "#/components/responses/Problem"
  /v1/oauth/keys/rotate:
    post:
      operationId: rotateSigningKey
      tags: [oauth-admin]
      summary: Activate the next token-signing key
      description: |
        The next key is the first key of `SIGNING_KEY_FILES` never
        activated. With the two-person rule (`KEY_ROTATION_TWO_PERSON`,
        default on) the first admin's call records the request (202)
        and a second admin's call within `KEY_ROTATION_CONFIRM_S`
        (600 s) activates it (200); the same admin twice is refused
        (409). The previous key is retired and stays in the JWKS for
        `KEY_RETIRE_GRACE_S` (24 h).
      security:
        - sessionCookie: []
      x-roles: [admin]
      x-audit: {event_type: signing_key_activated, entity_type: signing_key, purpose: none}
      responses:
        "200":
          description: The next key is active.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyRotation"
        "202":
          description: Requested; a second admin must confirm.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyRotation"
        default:
          $ref: "#/components/responses/Problem"

  # WP-3: the registry of operators, UAS and remote pilots (2019/947
  # Art. 14).
  # Registry records carry no personal data: an operator's or a pilot's
  # personal data is served only by the `personal-data` operations, to
  # the registrar and inspector roles, with a `purpose`, each read an
  # events row (`registry_pii_viewed`). A viewer, a USSP and the public
  # never reach them (CLAUDE.md rule 6).
...
  /v1/registry/uas/{uas_id}/status:
    parameters:
      - {name: uas_id, in: path, required: true, schema: {type: string, maxLength: 32}}
    post:
      operationId: setRegistryUASStatus
      tags: [registry]
      summary: Suspend, reinstate or revoke a UAS
      description: The transitions of setRegistryOperatorStatus.
      security:
        - sessionCookie: []
      x-roles: [registrar]
      x-audit: {event_type: registry_status_changed, entity_type: uas, purpose: none}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistryStatusInput"
      responses:
        "200":
          description: The aircraft after the change.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryUAS"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: uspace_session
      description: |
        Console session (decision M20): a JWT with `scope = "session"`,
        `roles[]`, `realm` (`console` or `police`), `jti`, `aud` = this
        system's host, set by the web BFF. State-changing requests also
        carry `X-CSRF-Token` matching the `uspace_csrf` cookie (M21).
    ecosystemToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Ecosystem machine token issued by a trusted issuer (RS256,
        `aud` = this system's host, scope from the catalogue in
        `docs/WORKPACKAGES/WP-2.md`), verified by uspace-core `auth`.
        Table B, the whole catalogue: `cis.read`, `cis.publish:zones`,
        `cis.publish:uspace`, `cis.publish:ussp_list`,
        `cis.publish:restrictions`, `cis.publish:ats_data` (reserved,
        never issued), `registry.validate`, `ussp.records`,
        `occurrences.write`, `certificates.status`, `police.query`,
        `ansp.traffic`, `ansp.coordination`, `ansp.requests`,
        `dp.observe` (lab-01 only), `utm.strategic_coordination`,
        `utm.constraint_processing`, `utm.constraint_management`,
        `utm.conformance_monitoring_sa`, `utm.availability_arbitration`,
        `rid.service_provider`, `rid.display_provider`, and at a USSP's
        issuer only `ussp.intents`, `ussp.telemetry`, `ussp.traffic`,
        `ussp.geo`.
    receiverKey:
      type: http
      scheme: bearer
      description: |
        Per-receiver bearer key of a Remote ID receiver, together with
        an HMAC-SHA256 of the exact body bytes under a separate secret
        (WP-7). Not a JWT and not a scope.
...